Fix visualization when detections do not provide a consistent shape property
Now the interactive visu uses the video shape. If no video is available, it computes the shape
from the detections and then from the track using the minimal shape to show most of the data.

When starting from frame_id != 0, we also update _video_frame so that the first shown frame is the correct one.