Use zstd to zip segmentation masks and recude memory usage.
Instead of storing the segmentation mask, we zip it and store the zip bytes in the tensor.
very efficient, without touching the code structure. Note that recent pytorch version (>2.6)
caches dropped segmentation tensor and re-allocate for the zip tensor (despite large fragmentation
in memory...). We found that using from_numpy instead of tensor avoid this phenomena allowing
to still benefits from the reduced memory usage. To be investigated.

This features is still experimental and at some point we may need to do a large rewrite of our Detection API
It can be triggers through ZSTD_SEG env/global variable.

We also add consistency for out of segmentation detections (NaN in pos, and 0 in bbox)